// frontend/src/app/api/models/categoria-guia-entrada.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CategoriaGuiaCampos } from '../models/categoria-guia-campos';
type CategoriaGuiaEntrada$ = CategoriaGuiaCampos;
type RequiredProperties = 'nombre' | 'slug' | 'descripcion' | 'orden' | 'activo';
export type CategoriaGuiaEntrada = CategoriaGuiaEntrada$ & Required<Pick<CategoriaGuiaEntrada$, RequiredProperties>>;