// frontend/src/app/api/models/config-inicio-entrada.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ConfigInicioCampos } from '../models/config-inicio-campos';
type ConfigInicioEntrada$ = ConfigInicioCampos;
type RequiredProperties = 'hero_titular' | 'hero_subtitulo' | 'hero_medio_id' | 'destinos_ids' | 'itinerarios_ids' | 'guias_ids';
export type ConfigInicioEntrada = ConfigInicioEntrada$ & Required<Pick<ConfigInicioEntrada$, RequiredProperties>>;