// frontend/src/app/api/models/configuracion-sitio-entrada.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ConfiguracionSitioCampos } from '../models/configuracion-sitio-campos';
type ConfiguracionSitioEntrada$ = ConfiguracionSitioCampos;
type RequiredProperties = 'nombre_marca' | 'texto_descargo';
export type ConfiguracionSitioEntrada = ConfiguracionSitioEntrada$ & Required<Pick<ConfiguracionSitioEntrada$, RequiredProperties>>;