// frontend/src/app/api/models/licencia-entrada.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { LicenciaCampos } from '../models/licencia-campos';
type LicenciaEntrada$ = LicenciaCampos;
type RequiredProperties = 'codigo' | 'nombre' | 'requiere_atribucion' | 'compatible_publicacion' | 'activo';
export type LicenciaEntrada = LicenciaEntrada$ & Required<Pick<LicenciaEntrada$, RequiredProperties>>;