// frontend/src/app/api/models/pais-entrada.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PaisCampos } from '../models/pais-campos';
type PaisEntrada$ = PaisCampos;
type RequiredProperties = 'nombre' | 'slug' | 'codigo_iso2' | 'region_id' | 'activo';
export type PaisEntrada = PaisEntrada$ & Required<Pick<PaisEntrada$, RequiredProperties>>;