// frontend/src/app/api/models/region-entrada.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { RegionCampos } from '../models/region-campos';
type RegionEntrada$ = RegionCampos;
type RequiredProperties = 'nombre' | 'slug' | 'continente' | 'orden' | 'activo';
export type RegionEntrada = RegionEntrada$ & Required<Pick<RegionEntrada$, RequiredProperties>>;